requirements.txt
python-telegram-bot==20.7
openai==1.3.7
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
